profile edit page and profile page separated

+++ models/profile.js
@@ -5,8 +5,8 @@ module.exports = function (sequelize, DataTypes) {
     var profile = sequelize.define("profile", {
         firstName: {type: DataTypes.STRING},
         lastName: {type: DataTypes.STRING},
-        phone: {type: DataTypes.INTEGER},
-        mobile: {type: DataTypes.INTEGER},
+        phone: {type: DataTypes.MEDIUMINT},
+        mobile: {type: DataTypes.MEDIUMINT},
         email: {type: DataTypes.STRING},
         gateKey: {type: DataTypes.STRING},
         bio: {type: DataTypes.STRING}
